+++ src/gui/Spells.java
@@ -19,8 +19,7 @@ import javax.swing.border.Border;
 
 @SuppressWarnings("serial")
 public class Spells extends JPanel {
-	private String[] attributes = {"spellName", "magicType", "magikaCost", "spellPotency",
-			   "duration", "prerequisite"};
+	private String[] attributes = {"spellName", "magicType", "magikaCost", "spellPotency", "duration", "prerequisite"};
 	
 	public Spells() throws Exception {
 		super();
@@ -90,29 +89,43 @@ public class Spells extends JPanel {
 			public void actionPerformed(final ActionEvent the_event) {
 				String searchBy = (String) searchByDropdown.getSelectedItem();
 				String condition = textField.getText();
-				StringBuilder query = new StringBuilder("SELECT ");
-				for (JCheckBox box : checkboxes) {
-					if (box.isSelected())
-						query.append(box.getText() + ", ");
+				boolean isInt = true;
+				try {
+					Integer.parseInt(condition);
+				} catch (NumberFormatException e) {
+					isInt = false;
 				}
-				if (query.substring(query.length() - 2, query.length()).equals(
-						", ")) {
-					query.delete(query.length() - 2, query.length());
-					query.append(" FROM `Spell` where " + searchBy
-							+ " = \"" + condition + "\";");
-					JTable table = null;
-					try {
-						table = Database.executeQuery(query.toString());
-					} catch (Exception e) {
-						e.printStackTrace();
+				if ((searchBy.equals("magikaCost") || searchBy.equals("spellPotency") || searchBy.equals("duration"))
+						&& !isInt || Integer.parseInt(condition) <= 0) {
+					JOptionPane.showMessageDialog(Spells.this, "Input must be a positive integer.",
+							"Input Error", JOptionPane.ERROR_MESSAGE);
+				} else {
+				
+					StringBuilder query = new StringBuilder("SELECT ");
+					for (JCheckBox box : checkboxes) {
+						if (box.isSelected())
+							query.append(box.getText() + ", ");
 					}
-					
-					JScrollPane scrollArea = new JScrollPane(table, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED, JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
-					table.setAutoResizeMode(JTable.AUTO_RESIZE_OFF);
-					resultsPanel.removeAll();
-					resultsPanel.add(scrollArea, BorderLayout.CENTER);
-					revalidate();
-					deleteButton.setEnabled(true);
+					if (query.substring(query.length() - 2, query.length()).equals(
+							", ")) {
+						query.delete(query.length() - 2, query.length());
+						query.append(" FROM `Spell` where " + searchBy
+								+ " = \"" + condition + "\";");
+						JTable table = null;
+						try {
+							table = Database.executeQuery(query.toString());
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
+						
+						JScrollPane scrollArea = new JScrollPane(table, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED,
+								JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
+						table.setAutoResizeMode(JTable.AUTO_RESIZE_OFF);
+						resultsPanel.removeAll();
+						resultsPanel.add(scrollArea, BorderLayout.CENTER);
+						revalidate();
+						deleteButton.setEnabled(true);
+					}
 				}
 			}
 		});
@@ -135,7 +148,8 @@ public class Spells extends JPanel {
 						e.printStackTrace();
 					}
 					
-					JScrollPane scrollArea = new JScrollPane(table, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED, JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
+					JScrollPane scrollArea = new JScrollPane(table, JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED,
+							JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
 					table.setAutoResizeMode(JTable.AUTO_RESIZE_OFF);
 					resultsPanel.removeAll();
 					resultsPanel.add(scrollArea, BorderLayout.CENTER);
